OldProjects/v2_1/requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Telegram Bot
aiogram>=3.4.1
aiohttp>=3.9.1

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
email-validator>=2.0.0

# Authentication & Security
pyjwt>=2.8.0
bcrypt>=4.1.1
keyring>=24.3.0
cryptography>=41.0.13

# Caching & Coordination
redis>=5.0.1
hiredis>=2.2.3

# Rate Limiting
slowapi>=0.1.9

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Excel (для импорта данных)
openpyxl>=3.1.2
msoffcrypto-tool>=5.3.0

# Active Directory
ldap3>=2.9.1

# Utilities
tabulate>=0.9.0
beautifulsoup4>=4.12.2
psutil>=5.9.6

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1



